Initial Project Commit
While this is a huge project commit, let me explain the contents within this update.

I began creating a project that would help me as a developer towards tracking web applications for a new job I'm applying for. This dashboard will be used to track projects, tasks, track everything in a calendar, and have the option of generating CSV and PDF project reports, and more. This project is primarily PHP with MySQL and other PHP dependencies. The goal, showcase my experience, competency, and understanding of PHP and alike systems for web application development.